pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zettelkasten-cli"
version = "0.1.0"
description = "A CLI tool to generate and manage a Zettelkasten knowledge base from podcast audio, videos, and articles"
authors = [{name = "Greg Weinger"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "anthropic>=0.18.0",
    "openai-whisper>=20231117",
    "feedparser>=6.0.0",
    "yt-dlp>=2023.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
zk = "zettelkasten.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["zettelkasten*"]
exclude = ["vault*", "downloads*", "transcripts*", "tests*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
